% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bene}
\alias{bene}
\title{CMS Beneficiary Summary DE-SynPUF}
\format{
A data frame with 998 rows and 33 variables:
\describe{
\item{\code{BENE_YEAR}}{Beneficiary Enrollment Year}
\item{\code{DESYNPUF_ID}}{Beneficiary Code}
\item{\code{BENE_BIRTH_DT}}{Date of birth}
\item{\code{BENE_DEATH_DT}}{Date of death}
\item{\code{BENE_SEX_IDENT_CD}}{Sex}
\item{\code{BENE_RACE_CD}}{Beneficiary Race Code}
\item{\code{BENE_ESRD_IND}}{End stage renal disease Indicator}
\item{\code{SP_STATE_CODE}}{State Code}
\item{\code{BENE_COUNTY_CD}}{County Code}
\item{\code{BENE_HI_CVRAGE_TOT_MONS}}{Total number of months of part A coverage for the beneficiary}
\item{\code{BENE_SMI_CVRAGE_TOT_MONS}}{Total number of months of part B coverage for the beneficiary}
\item{\code{BENE_HMO_CVRAGE_TOT_MONS}}{Total number of months of HMO coverage for the beneficiary}
\item{\code{PLAN_CVRG_MOS_NUM}}{Total number of months of part D plan coverage for the beneficiary}
\item{\code{SP_ALZHDMTA}}{Chronic Condition Alzheimer or related disorders or senile}
\item{\code{SP_CHF}}{Chronic Condition Heart Failure}
\item{\code{SP_CHRNKIDN}}{Chronic Condition Chronic Kidney Disease}
\item{\code{SP_CNCR}}{Chronic Condition Cancer}
\item{\code{SP_COPD}}{Chronic Condition Chronic Obstructive Pulmonary Disease}
\item{\code{SP_DEPRESSN}}{Chronic Condition Depression}
\item{\code{SP_DIABETES}}{Chronic Condition Diabetes}
\item{\code{SP_ISCHMCHT}}{Chronic Condition Ischemic Heart Disease}
\item{\code{SP_OSTEOPRS}}{Chronic Condition Osteoporosis}
\item{\code{SP_RA_OA}}{Chronic Condition RAOA}
\item{\code{SP_STRKETIA}}{Chronic Condition Stroketransient Ischemic Attack}
\item{\code{MEDREIMB_IP}}{Inpatient annual Medicare reimbursement amount}
\item{\code{BENRES_IP}}{Inpatient annual beneficiary responsibility amount}
\item{\code{PPPYMT_IP}}{Inpatient annual primary payer reimbursement amount}
\item{\code{MEDREIMB_OP}}{Outpatient Institutional annual Medicare reimbursement amount}
\item{\code{BENRES_OP}}{Outpatient Institutional annual beneficiary responsibility amount}
\item{\code{PPPYMT_OP}}{Outpatient Institutional annual primary payer reimbursement amount}
\item{\code{MEDREIMB_CAR}}{Carrier annual Medicare reimbursement amount}
\item{\code{BENRES_CAR}}{Carrier annual beneficiary responsibility amount}
\item{\code{PPPYMT_CAR}}{Carrier annual primary payer reimbursement amount}
}
}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
bene
}
\description{
Synthetic Medicare beneficiary records
}
\keyword{datasets}
