% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpatient}
\alias{inpatient}
\title{CMS Inpatient Claims DE-SynPUF}
\format{
A data frame with 225 rows and 35 variables:
\describe{
\item{\code{DESYNPUF_ID}}{Beneficiary Code}
\item{\code{CLM_ID}}{Claim ID}
\item{\code{SEGMENT}}{Claim Line Segment}
\item{\code{CLM_FROM_DT}}{Claims start date}
\item{\code{CLM_THRU_DT}}{Claims end date}
\item{\code{PRVDR_NUM}}{Provider Institution}
\item{\code{CLM_PMT_AMT}}{Claim Payment Amount}
\item{\code{NCH_PRMRY_PYR_CLM_PD_AMT}}{NCH Primary Payer Claim Paid Amount}
\item{\code{AT_PHYSN_NPI}}{Attending Physician National Provider Identifier Number}
\item{\code{OP_PHYSN_NPI}}{Operating Physician National Provider Identifier Number}
\item{\code{OT_PHYSN_NPI}}{Other Physician National Provider Identifier Number}
\item{\code{CLM_ADMSN_DT}}{Inpatient admission date}
\item{\code{ADMTNG_ICD9_DGNS_CD}}{Claim Admitting Diagnosis Code}
\item{\code{CLM_PASS_THRU_PER_DIEM_AMT}}{Claim Pass Thru Per Diem Amount}
\item{\code{NCH_BENE_IP_DDCTBL_AMT}}{NCH Beneficiary Inpatient Deductible Amount}
\item{\code{NCH_BENE_PTA_COINSRNC_LBLTY_AM}}{NCH Beneficiary Part A Coinsurance Liability Amount}
\item{\code{NCH_BENE_BLOOD_DDCTBL_LBLTY_AM}}{NCH Beneficiary Blood Deductible Liability Amount}
\item{\code{CLM_UTLZTN_DAY_CNT}}{Claim Utilization Day Count}
\item{\code{NCH_BENE_DSCHRG_DT}}{Inpatient discharged date}
\item{\code{CLM_DRG_CD}}{Claim Diagnosis Related Group Code}
\item{\code{ICD9_DGNS_CD_1}}{Claim Diagnosis Code (1)}
\item{\code{ICD9_DGNS_CD_2}}{Claim Diagnosis Code (2)}
\item{\code{ICD9_DGNS_CD_3}}{Claim Diagnosis Code (3)}
\item{\code{ICD9_DGNS_CD_4}}{Claim Diagnosis Code (4)}
\item{\code{ICD9_DGNS_CD_5}}{Claim Diagnosis Code (5)}
\item{\code{ICD9_PRCDR_CD_1}}{Claim Procedure Code (1)}
\item{\code{ICD9_PRCDR_CD_2}}{Claim Procedure Code (2)}
\item{\code{ICD9_PRCDR_CD_3}}{Claim Procedure Code (3)}
\item{\code{ICD9_PRCDR_CD_4}}{Claim Procedure Code (4)}
\item{\code{ICD9_PRCDR_CD_5}}{Claim Procedure Code (5)}
\item{\code{HCPCS_CD_1}}{Revenue Center HCFA Common Procedure Coding System (1)}
\item{\code{HCPCS_CD_2}}{Revenue Center HCFA Common Procedure Coding System (2)}
\item{\code{HCPCS_CD_3}}{Revenue Center HCFA Common Procedure Coding System (3)}
\item{\code{HCPCS_CD_4}}{Revenue Center HCFA Common Procedure Coding System (4)}
\item{\code{HCPCS_CD_5}}{Revenue Center HCFA Common Procedure Coding System (5)}
}
}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
inpatient
}
\description{
Synthetic inpatient claim records
}
\keyword{datasets}
