% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pde}
\alias{pde}
\title{CMS Prescription Drug Events (PDE) DE-SynPUF}
\format{
A data frame with 17,443 rows and 8 variables:
\describe{
\item{\code{DESYNPUF_ID}}{Beneficiary Code}
\item{\code{PDE_ID}}{CCW Part D Event Number}
\item{\code{SRVC_DT}}{RX Service Date}
\item{\code{PROD_SRVC_ID}}{Product Service ID}
\item{\code{QTY_DSPNSD_NUM}}{Quantity Dispensed}
\item{\code{DAYS_SUPLY_NUM}}{Days Supply}
\item{\code{PTNT_PAY_AMT}}{Patient Pay Amount}
\item{\code{TOT_RX_CST_AMT}}{Gross Drug Cost}
}
}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}
}
\usage{
pde
}
\description{
Synthetic Part D event records
}
\keyword{datasets}
