% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schema}
\alias{schema}
\title{CMS Data Entrepreneurs’ Synthetic Public Use File Schema}
\format{
A data frame with 5 rows and 5 variables:
\describe{
\item{\code{TABLE}}{Name of the DE_SynPUF table}
\item{\code{TABLE_TITLE}}{Full title of the DE_SynPUF table}
\item{\code{TABLE_DESCRIPTION}}{Description of the DE_SynPUF table}
\item{\code{UNIT_OF_RECORD}}{Description of the type of records in the table}
\item{\code{PROPERTIES}}{The properties of the table as a nested table containing the \code{VARIABLE}, the data \code{TYPE}, and a \code{DESCRIPTION} of the columns}
}
}
\usage{
schema
}
\description{
Includes information about the schema of the tables that were sourced from
\href{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DE_Syn_PUF}{CMS},
the \emph{CMS 2008-2010 Data Entrepreneurs’ Synthetic Public Use File (DE-SynPUF)}.
This package include 500 randomly selected beneficiaries from the 2008 file in
sample 2, along with their associated beneficiary records and claims for 2008
and 2009. Not all variables from the DE-SynPUF are columns in the data
in this package: numbered fields (e.g., claims diagnosis codes) were limited
to the first 5 to limit file size. Full documentation for the DE-SynPUF is
available for sample 2 from \href{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SynPUFs/DESample02}{CMS}.
}
\keyword{datasets}
